updated css to change font white
