+++ src/app/containers/trips/inDepthDisplayTrip.tsx
@@ -27,7 +27,7 @@ export default function InDepthDisplayTrip({
                 <p>Carrier: {trip?.carrier}<br />
                     Carrier Address: {trip?.carrieraddress}<br />
                     Inspection Address: {trip?.inspectionaddress}<br/>
-                    Inspection Time: {trip?.inputdate.toString()}
+                    Inspection Time: {trip?.inputdate.toLocaleString()}
                 </p>
             </div>
             <div className={styles.inDepthTripInfo}>
